🚛 continue pulling resources into components/
